LPS-59471 Move getAssetEntry and addAssetLink methods to BaseExportImportTestCase
